Change name and variants
cayo -> spamtagger-bootc
centos -> spamtagger
fedora -> spamtagger-plus

This makes for somewhat clumsy versions like:

  spamtagger-bootc-spamtagger-plus-10

but this can be ignored with simpler tag aliases.

The `gen-tags` function in the Justfile looks for existing images in the
registry, so we need to spoof those before the first upload.

Both variants are now based off of centos, but there are likely other
fedora-specific sections which need to be removed still.

Nothing else has yet changed, so both variants should essentially be
equivalent to cayo. The next steps will be to remove things added to
cayo that we don't need, then start adding things that we do need.