Improve Anaconda ISO builds, fix disk formats
Try adding necessary files to build a custom 'product.img' to override
Anaconda defaults. This doesn't appear to work correctly. Seems that
'product.img' is not actually taken from the working directory of the
parent shell (running the build via 'podman'), or the 'product.img' is
invalid. I suspect the former.

Change bundle outputs to build/bundles

Change 'ami' and 'ova' '$diskformat' to 'vmdk'.

Change from 7z to zip. Disk images are already fairly compressed, so 7z
doesn't help that much. Zip is more ubiquitous, and is what MailCleaner
used, so it is a better bet.