Change build-disk method
I was having trouble actually getting bootc-image-builder to work. It
appears to be fairly tightly integrated with RPM tools, so a different
(simpler) method will have to do.

This change just creates a virtual disk image, mounts it to a new
container and uses Bootc on that container to write to the virtual disk.

This works insofar as it builds seems to copy the appropriate bytes to
the image and then that image can be converted to VM disk formats. It
does not work insofar as the disk images are not bootable. I think this
is an issue in `build-container` not `build-disk`. We probably don't
have the bootloader configured properly. To be investigated next.